WIP Simplify SideDrawer --> Reorganize drawer stuff
